+++ app/events/event.tsx
@@ -6,21 +6,18 @@ import {
   useTheme,
   useMediaQuery,
   Stack,
-  Grid,
   OutlinedInput,
   InputAdornment,
-  Button,
   Skeleton,
 } from '@mui/material';
 import debounce from 'lodash/debounce';
 import { Sidebar } from 'components/layout';
 import SidebarLeft from './components/Sidebar';
-import { EventCard, LotteryCard } from '@/components/cards';
-// import SelectButton from '@/components/buttons/SelectButton';
+import { EventCard } from '@/components/cards';
 import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
 import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
 import { useCeramicContext } from '../../context/CeramicContext';
-import { Event, EventData, LegacyEvent } from '@/types';
+import { Event } from '@/types';
 import { EventIcon, SearchIcon } from '@/components/icons';
 import EventHeader from './components/EventHeader';
 import {
@@ -30,6 +27,7 @@ import {
 } from '@/components/cards/EventCard';
 import { supabase } from '@/utils/supabase/client';
 import dayjs from 'dayjs';
+import EventList from '@/components/event/EventList';
 
 const EventsPage: React.FC = () => {
   const theme = useTheme();
@@ -187,7 +185,6 @@ const EventsPage: React.FC = () => {
     return groupedEvents;
   }, [events]);
 
-  // TODO: Implement search functionality
   const onSearch = () => {
     getEvents()
       .catch((error) => console.error('Failed to fetch events:', error))
@@ -215,57 +212,14 @@ const EventsPage: React.FC = () => {
         {!isTablet && <Sidebar selected={selected} />}
         <Stack direction="column" borderLeft="1px solid #383838" flex={1}>
           <EventHeader />
-          <Box
-            display="flex"
-            padding={'0 20px'}
-            margin={'20px 0 0 0'}
-            justifyContent="space-between"
-          >
-            <Box display="flex" alignItems="center" gap="10px">
-              <EventIcon />
-              <Typography color="white" variant="subtitleLB">
-                Events
-              </Typography>
-            </Box>
-            {/*<Box display={"flex"} position={"relative"} justifyContent={"end"}>*/}
-            {/*    <SelectButton Icon={CalendarIcon} title={'Dates'} onOpened={() => setShowDate(true)} onClosed={() => setShowDate(false)} />*/}
-            {/*    {showDate && <Box backgroundColor={"#222"} position={"absolute"} top={"40px"} padding={"10px"} borderRadius={"10px"} zIndex={"1000"} right={"0"}>*/}
-            {/*        <Box*/}
-            {/*            display="flex"*/}
-            {/*            gap="4px"*/}
-            {/*            padding="2px"*/}
-            {/*            borderRadius="10px"*/}
-            {/*            bgcolor="#2d2d2d"*/}
-            {/*        >*/}
-            {/*            <Button*/}
-            {/*                sx={{*/}
-            {/*                    flex: 1,*/}
-            {/*                    backgroundColor: '#424242',*/}
-            {/*                    borderRadius: '8px',*/}
-            {/*                    color: 'white',*/}
-            {/*                    fontFamily: 'Inter',*/}
-            {/*                }}*/}
-            {/*            >*/}
-            {/*                Upcoming*/}
-            {/*            </Button>*/}
-            {/*            <Button*/}
-            {/*                sx={{*/}
-            {/*                    flex: 1,*/}
-            {/*                    backgroundColor: '#2d2d2d',*/}
-            {/*                    borderRadius: '8px',*/}
-            {/*                    color: 'white',*/}
-            {/*                    fontFamily: 'Inter',*/}
-            {/*                }}*/}
-            {/*            >*/}
-            {/*                Past*/}
-            {/*            </Button>*/}
-            {/*        </Box>*/}
-            {/*        <Box>*/}
-            {/*            <ZuCalendar defaultValue={dayjs('2022-04-17')}/>*/}
-            {/*        </Box>*/}
-            {/*    </Box>}*/}
-            {/*</Box>*/}
-          </Box>
+          <Stack p="20px">
+            <EventList
+              top={50}
+              hasAllButton={false}
+              events={events}
+              isLoading={isEventsLoading}
+            />
+          </Stack>
           <Stack
             sx={{
               display: 'none',
@@ -307,48 +261,6 @@ const EventsPage: React.FC = () => {
                 </InputAdornment>
               }
             />
-          </Stack>
-          <Stack
-            sx={{
-              display: 'flex',
-              flexDirection: 'column',
-              flexWrap: 'wrap',
-              gap: '20px',
-              padding: '20px',
-              justifyContent: 'center',
-            }}
-          >
-            {isEventsLoading ? (
-              <>
-                <EventCardMonthGroup>
-                  <Skeleton width={60}></Skeleton>
-                </EventCardMonthGroup>
-                <EventCardSkeleton />
-                <EventCardSkeleton />
-              </>
-            ) : (
-              events.length > 0 &&
-              Object.entries(eventsData).map(([month, events], index) => {
-                return (
-                  <Fragment key={month + index}>
-                    <EventCardMonthGroup>{month}</EventCardMonthGroup>
-                    {events.map((event) => (
-                      <EventCard key={event.id} event={event} />
-                    ))}
-                    {/*<Grid*/}
-                    {/*  key={`Lottery-Card`}*/}
-                    {/*  xs={12}*/}
-                    {/*  sm={6}*/}
-                    {/*  md={4}*/}
-                    {/*  xl={3}*/}
-                    {/*  sx={{ display: 'flex', justifyContent: 'center' }}*/}
-                    {/*>*/}
-                    {/*  <LotteryCard />*/}
-                    {/*</Grid>*/}
-                  </Fragment>
-                );
-              })
-            )}
           </Stack>
         </Stack>
         <SidebarLeft
